+++ .buildkite/buildkite-shared/buildkite_shared/uv.py
@@ -1,0 +1,1 @@
+UV_PIN = "uv==0.7.2"
