+++ .buildkite/dagster-buildkite/dagster_buildkite/cli.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from dagster_buildkite.git import GitInfo
+from buildkite_shared.git import GitInfo
 from dagster_buildkite.pipelines.dagster_oss_main import build_dagster_oss_main_steps
 from dagster_buildkite.pipelines.dagster_oss_nightly_pipeline import (
     build_dagster_oss_nightly_steps,
@@ -8,7 +8,7 @@ from dagster_buildkite.pipelines.dagster_oss_nightly_pipeline import (
 from dagster_buildkite.pipelines.prerelease_package import (
     build_prerelease_package_steps,
 )
-from dagster_buildkite.python_packages import PythonPackages
+from buildkite_shared.python_packages import PythonPackages
 from dagster_buildkite.utils import buildkite_yaml_for_steps
 
 CLI_HELP = """This CLI is used for generating Buildkite YAML. Each function corresponds to an entry
