+++ .buildkite/dagster-buildkite/dagster_buildkite/steps/helm.py
@@ -1,7 +1,7 @@
 import os
 from typing import List
 
-from dagster_buildkite.package_spec import PackageSpec
+from dagster_buildkite.steps.packages import PackageSpec
 from buildkite_shared.python_version import AvailablePythonVersion
 from buildkite_shared.step_builders.command_step_builder import (
     CommandStepBuilder,
